+++ process_data/parse_train.py
@@ -143,7 +143,7 @@ parsed['sender_state'] = dis_attr[4]
 parsed['receive_state'] = dis_attr[5]
 
 parsed['acc_hour'] = fattr[1]
-parsed['isNextDay'] = parsed['acc_hour'] >= 14
+parsed['isNextDay'] = (parsed['acc_hour'] >= 14).astype(int)
 parsed['pay_hour'] = fattr[4]
 parsed['acc_date'] = fattr[5]
 parsed['isHoliday'] = fattr[6]
